requirements.txt
google-generativeai==0.8.3
streamlit==1.43.0
python-dotenv==1.0.1
db-sqlite3==0.0.1
pandas==2.2.3